requirements.txt
biopython>=1.83
py3Dmol>=2.1.0
matplotlib>=3.8.0
numpy>=1.26.0
requests>=2.31.0
Pillow>=10.0.0
seaborn>=0.13.0
